Refactor unit tests for route and virtual tests
