+++ tests/syft/core/io/route_test.py
@@ -1,17 +1,6 @@
-# stdlib
-from typing import Callable
-from typing import Union
-
-# third party
-from nacl.bindings.crypto_sign import crypto_sign_keypair
-from nacl.signing import VerifyKey
-
-# syft absolute
 from syft.core.common.message import SignedEventualSyftMessageWithoutReply
 from syft.core.common.message import SignedImmediateSyftMessageWithReply
 from syft.core.common.message import SignedImmediateSyftMessageWithoutReply
-from syft.core.common.uid import UID
-from syft.core.io.address import Address
 from syft.core.io.location.specific import SpecificLocation
 from syft.core.io.route import BroadcastRoute
 from syft.core.io.route import Route
@@ -20,63 +9,8 @@ from syft.core.io.route import SoloRoute
 from syft.core.io.virtual import VirtualClientConnection
 from syft.core.io.virtual import VirtualServerConnection
 from syft.core.node.common.node import Node
-
-# --------------------- TEST HELPERS ---------------------
-
-
-class MockNode(Node):
-    """Mock Node object for testing purposes."""
-
-    def recv_immediate_msg_without_reply(
-        self, msg: SignedImmediateSyftMessageWithoutReply
-    ) -> None:
-        return None
-
-    def recv_eventual_msg_without_reply(
-        self, msg: SignedImmediateSyftMessageWithReply
-    ) -> None:
-        return None
-
-    def recv_immediate_msg_with_reply(
-        self, msg: SignedImmediateSyftMessageWithReply
-    ) -> SignedImmediateSyftMessageWithoutReply:
-        return _construct_dummy_message(SignedImmediateSyftMessageWithoutReply)
-
-
-def _construct_address() -> Address:
-    """Helper method to construct an Address"""
-    return Address(
-        network=SpecificLocation(id=UID()),
-        domain=SpecificLocation(id=UID()),
-        device=SpecificLocation(id=UID()),
-        vm=SpecificLocation(id=UID()),
-    )
-
-
-def _construct_dummy_message(
-    msg_class: Callable,
-) -> Union[
-    SignedEventualSyftMessageWithoutReply,
-    SignedImmediateSyftMessageWithReply,
-    SignedImmediateSyftMessageWithoutReply,
-]:
-    """
-    Helper method to construct a dummy SignedMessage of the following:
-    - SignedEventualSyftMessageWithoutReply
-    - SignedImmediateSyftMessageWithReply
-    - SignedImmediateSyftMessageWithoutReply
-    """
-    address = _construct_address()
-    key = VerifyKey(crypto_sign_keypair()[0])
-
-    return msg_class(
-        address=address,
-        obj_type="signed_message",
-        signature=b"my_signature",
-        verify_key=key,
-        message=b"hello_world",
-    )
-
+from tests.syft.core.io.utils_test import construct_dummy_message
+from tests.syft.core.io.utils_test import MockNode
 
 # --------------------- INITIALIZATION ---------------------
 
@@ -151,7 +85,7 @@ def test_solo_route_send_immediate_msg_without_reply() -> None:
     connection = VirtualClientConnection(server=server)
 
     h_solo = SoloRoute(destination=destination, connection=connection)
-    msg = _construct_dummy_message(SignedImmediateSyftMessageWithoutReply)
+    msg = construct_dummy_message(SignedImmediateSyftMessageWithoutReply)
 
     assert h_solo.send_immediate_msg_without_reply(msg) is None
 
@@ -164,7 +98,7 @@ def test_solo_route_send_eventual_msg_without_reply() -> None:
     connection = VirtualClientConnection(server=server)
 
     h_solo = SoloRoute(destination=destination, connection=connection)
-    msg = _construct_dummy_message(SignedEventualSyftMessageWithoutReply)
+    msg = construct_dummy_message(SignedEventualSyftMessageWithoutReply)
 
     assert not h_solo.send_eventual_msg_without_reply(msg)
 
@@ -177,7 +111,7 @@ def test_solo_route_send_immediate_msg_with_reply() -> None:
     connection = VirtualClientConnection(server=server)
 
     h_solo = SoloRoute(destination=destination, connection=connection)
-    msg = _construct_dummy_message(SignedImmediateSyftMessageWithReply)
+    msg = construct_dummy_message(SignedImmediateSyftMessageWithReply)
     ret = h_solo.send_immediate_msg_with_reply(msg)
 
     assert isinstance(
